+++ src/axom/mint/mesh/internal/ConnectivityArray_indirection.hpp
@@ -313,8 +313,8 @@ public:
   {
     m_offsets->resize( ID_size+1 );
     IndexType newValueSize =
-        internal::calcValueCapacity( getNumberOfIDs(), getIDCapacity(),
-                                     getNumberOfValues(), value_size );
+      internal::calcValueCapacity( getNumberOfIDs(), getIDCapacity(),
+                                   getNumberOfValues(), value_size );
     m_values->resize( newValueSize );
   }