+++ src/axom/sidre/interface/c_fortran/wrapDataStore.cpp
@@ -109,7 +109,7 @@ SIDRE_buffer* SIDRE_datastore_create_buffer_empty(SIDRE_datastore* self,
 
 SIDRE_buffer* SIDRE_datastore_create_buffer_from_type(SIDRE_datastore* self,
                                                       int type,
-                                                      SIDRE_SidreLength num_elems,
+                                                      SIDRE_IndexType num_elems,
                                                       SIDRE_buffer* SHC_rv)
 {
 // splicer begin class.DataStore.method.create_buffer_from_type