+++ src/axom/sidre/interface/c_fortran/wrapGroup.cpp
@@ -398,7 +398,7 @@ SIDRE_view* SIDRE_group_create_view_empty_bufferify(SIDRE_group* self,
 
 SIDRE_view* SIDRE_group_create_view_from_type(SIDRE_group* self,
                                               const char* path, int type,
-                                              SIDRE_SidreLength num_elems,
+                                              SIDRE_IndexType num_elems,
                                               SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_type
@@ -424,7 +424,7 @@ SIDRE_view* SIDRE_group_create_view_from_type(SIDRE_group* self,
 SIDRE_view* SIDRE_group_create_view_from_type_bufferify(SIDRE_group* self,
                                                         const char* path,
                                                         int Lpath, int type,
-                                                        SIDRE_SidreLength num_elems,
+                                                        SIDRE_IndexType num_elems,
                                                         SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_type_bufferify
@@ -442,7 +442,7 @@ SIDRE_view* SIDRE_group_create_view_from_type_bufferify(SIDRE_group* self,
 SIDRE_view* SIDRE_group_create_view_from_shape(SIDRE_group* self,
                                                const char* path, int type,
                                                int ndims,
-                                               SIDRE_SidreLength* shape,
+                                               SIDRE_IndexType* shape,
                                                SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_shape
@@ -469,7 +469,7 @@ SIDRE_view* SIDRE_group_create_view_from_shape_bufferify(SIDRE_group* self,
                                                          const char* path,
                                                          int Lpath, int type,
                                                          int ndims,
-                                                         SIDRE_SidreLength* shape,
+                                                         SIDRE_IndexType* shape,
                                                          SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_shape_bufferify
@@ -530,7 +530,7 @@ SIDRE_view* SIDRE_group_create_view_into_buffer_bufferify(SIDRE_group* self,
 SIDRE_view* SIDRE_group_create_view_from_type_and_buffer(SIDRE_group* self,
                                                          const char* path,
                                                          int type,
-                                                         SIDRE_SidreLength num_elems, SIDRE_buffer* buff,
+                                                         SIDRE_IndexType num_elems, SIDRE_buffer* buff,
                                                          SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_type_and_buffer
@@ -557,7 +557,7 @@ SIDRE_view* SIDRE_group_create_view_from_type_and_buffer(SIDRE_group* self,
 
 SIDRE_view* SIDRE_group_create_view_from_type_and_buffer_bufferify(
   SIDRE_group* self, const char* path, int Lpath, int type,
-  SIDRE_SidreLength num_elems, SIDRE_buffer* buff, SIDRE_view* SHC_rv)
+  SIDRE_IndexType num_elems, SIDRE_buffer* buff, SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_type_and_buffer_bufferify
   axom::sidre::Group* SH_this = static_cast<axom::sidre::Group*>(self->addr);
@@ -576,7 +576,7 @@ SIDRE_view* SIDRE_group_create_view_from_type_and_buffer_bufferify(
 SIDRE_view* SIDRE_group_create_view_from_shape_and_buffer(SIDRE_group* self,
                                                           const char* path,
                                                           int type, int ndims,
-                                                          SIDRE_SidreLength* shape, SIDRE_buffer* buff,
+                                                          SIDRE_IndexType* shape, SIDRE_buffer* buff,
                                                           SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_shape_and_buffer
@@ -603,7 +603,7 @@ SIDRE_view* SIDRE_group_create_view_from_shape_and_buffer(SIDRE_group* self,
 
 SIDRE_view* SIDRE_group_create_view_from_shape_and_buffer_bufferify(
   SIDRE_group* self, const char* path, int Lpath, int type, int ndims,
-  SIDRE_SidreLength* shape, SIDRE_buffer* buff, SIDRE_view* SHC_rv)
+  SIDRE_IndexType* shape, SIDRE_buffer* buff, SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_shape_and_buffer_bufferify
   axom::sidre::Group* SH_this = static_cast<axom::sidre::Group*>(self->addr);
@@ -661,7 +661,7 @@ SIDRE_view* SIDRE_group_create_view_external_bufferify(SIDRE_group* self,
 SIDRE_view* SIDRE_group_create_view_from_type_external(SIDRE_group* self,
                                                        const char* path,
                                                        int type,
-                                                       SIDRE_SidreLength num_elems, void* external_ptr,
+                                                       SIDRE_IndexType num_elems, void* external_ptr,
                                                        SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_type_external
@@ -686,7 +686,7 @@ SIDRE_view* SIDRE_group_create_view_from_type_external(SIDRE_group* self,
 
 SIDRE_view* SIDRE_group_create_view_from_type_external_bufferify(
   SIDRE_group* self, const char* path, int Lpath, int type,
-  SIDRE_SidreLength num_elems, void* external_ptr, SIDRE_view* SHC_rv)
+  SIDRE_IndexType num_elems, void* external_ptr, SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_type_external_bufferify
   axom::sidre::Group* SH_this = static_cast<axom::sidre::Group*>(self->addr);
@@ -703,7 +703,8 @@ SIDRE_view* SIDRE_group_create_view_from_type_external_bufferify(
 SIDRE_view* SIDRE_group_create_view_from_shape_external(SIDRE_group* self,
                                                         const char* path,
                                                         int type, int ndims,
-                                                        SIDRE_SidreLength* shape, void* external_ptr,
+                                                        SIDRE_IndexType* shape,
+                                                        void* external_ptr,
                                                         SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_shape_external
@@ -728,7 +729,7 @@ SIDRE_view* SIDRE_group_create_view_from_shape_external(SIDRE_group* self,
 
 SIDRE_view* SIDRE_group_create_view_from_shape_external_bufferify(
   SIDRE_group* self, const char* path, int Lpath, int type, int ndims,
-  SIDRE_SidreLength* shape, void* external_ptr, SIDRE_view* SHC_rv)
+  SIDRE_IndexType* shape, void* external_ptr, SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_from_shape_external_bufferify
   axom::sidre::Group* SH_this = static_cast<axom::sidre::Group*>(self->addr);
@@ -745,7 +746,7 @@ SIDRE_view* SIDRE_group_create_view_from_shape_external_bufferify(
 SIDRE_view* SIDRE_group_create_view_and_allocate_nelems(SIDRE_group* self,
                                                         const char* path,
                                                         int type,
-                                                        SIDRE_SidreLength num_elems,
+                                                        SIDRE_IndexType num_elems,
                                                         SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_and_allocate_nelems
@@ -771,7 +772,7 @@ SIDRE_view* SIDRE_group_create_view_and_allocate_nelems(SIDRE_group* self,
 
 SIDRE_view* SIDRE_group_create_view_and_allocate_nelems_bufferify(
   SIDRE_group* self, const char* path, int Lpath, int type,
-  SIDRE_SidreLength num_elems, SIDRE_view* SHC_rv)
+  SIDRE_IndexType num_elems, SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_and_allocate_nelems_bufferify
   axom::sidre::Group* SH_this = static_cast<axom::sidre::Group*>(self->addr);
@@ -789,7 +790,7 @@ SIDRE_view* SIDRE_group_create_view_and_allocate_nelems_bufferify(
 SIDRE_view* SIDRE_group_create_view_and_allocate_shape(SIDRE_group* self,
                                                        const char* path,
                                                        int type, int ndims,
-                                                       SIDRE_SidreLength* shape,
+                                                       SIDRE_IndexType* shape,
                                                        SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_and_allocate_shape
@@ -815,7 +816,7 @@ SIDRE_view* SIDRE_group_create_view_and_allocate_shape(SIDRE_group* self,
 
 SIDRE_view* SIDRE_group_create_view_and_allocate_shape_bufferify(
   SIDRE_group* self, const char* path, int Lpath, int type, int ndims,
-  SIDRE_SidreLength* shape, SIDRE_view* SHC_rv)
+  SIDRE_IndexType* shape, SIDRE_view* SHC_rv)
 {
 // splicer begin class.Group.method.create_view_and_allocate_shape_bufferify
   axom::sidre::Group* SH_this = static_cast<axom::sidre::Group*>(self->addr);